Add arrow to overlay tip pointing where to look for offscreen elements
This change builds on top of the overlay rework for multiple highlighted rectangles:
https://github.com/bvaughn/react-devtools-experimental/pull/207

If the highlighted elements are offscreen, the tooltip is positioned
towards these elements, but the element highlights (OverlayRect)
are not visible, so if the element size is zero or close to it,
it's hard to visually tell apart an onscreen element located where
the tooltip is positioned and an offscreen element located elsewhere.

This change improves the UX by adding to the tooltips
a little arrow pointing to where the elements are located
if they are offscreen.

For onscreen elements the tooltip does not change.