Implement showing native elements under a persisted Settings item
It's often necessary to debug native (host) elements rendered by
the components, but the new version of DevTools removed this functionality.

This change restores the functionality under a persisted setting,
working similar to "reload and profile", so that by default
the performance of the DevTools is unaffected.

To distinguish the HostComponent and HostText types in Components,
new ElementType items were added, and new CSS variables were added
for colors in both light and dart themes.

Within this diff there's also a fix to the shells/dev iframe reload handling.
Previously reloading the iframe would not restore the test app
and would break the bridge/store. This diff more accurately resembles
the load/unload/messaging behavior and the code written for the extension.

The reasoning for this feature to exist:

> I want to do the reverse: from the tree find where are my elements.
> Not every element that gets rendered is visible and of clickable size.
> https://twitter.com/sompylasar/status/1118580156430311424
> https://twitter.com/dan_abramov/status/1118486315736158208