Allow Inspecting Elements Within Iframes
This change adds support for element inspecting within `<iframe/>`s.

The iframes are kept in a Set so that we only append listeners once and
clean them up properly. I’ve run a few tests and it seems to behave as
expected.

Here’s a screencapture: http://cl.ly/358a8f22715d/iframe-demo.mov

The fixture includes a cross-origin iframe to make sure we do not error
in this case.

I’ve run through the following scenarios locally:

- Nexted iframes
- Iframe is removed while we’re inspecting
- Chrome and Firefox